refactor: enum type 에 대한 package 생성
